+++ src/js/base/api.js
@@ -1,22 +1,9 @@
-if (
-    typeof require !== "undefined" &&
-    (typeof window === "undefined" ||
-        // eslint-disable-next-line camelcase
-        typeof __webpack_require__ !== "undefined" ||
-        (typeof navigator !== "undefined" && navigator.product === "ReactNative"))
-) {
-    // eslint-disable-next-line no-redeclare,no-var
-    var base = require("./base");
-    // eslint-disable-next-line no-redeclare,no-var
-    var compat = require("./compat");
-    require("./ripe");
-    // eslint-disable-next-line no-redeclare,no-var
-    var ripe = base.ripe;
-    // eslint-disable-next-line no-redeclare,no-var
-    var fetch = compat.fetch;
-    // eslint-disable-next-line no-redeclare,no-var
-    var XMLHttpRequest = compat.XMLHttpRequest;
-}
+const base = require("./base");
+const compat = require("./compat");
+require("./ripe");
+const ripe = base.ripe;
+const fetch = compat.fetch;
+const XMLHttpRequest = compat.XMLHttpRequest;
 
 /**
  * @class