+++ src/js/base/config.js
@@ -1,16 +1,6 @@
-if (
-    typeof require !== "undefined" &&
-    (typeof window === "undefined" ||
-        // eslint-disable-next-line camelcase
-        typeof __webpack_require__ !== "undefined" ||
-        (typeof navigator !== "undefined" && navigator.product === "ReactNative"))
-) {
-    // eslint-disable-next-line no-redeclare,no-var
-    var base = require("./base");
-    require("./ripe");
-    // eslint-disable-next-line no-redeclare,no-var
-    var ripe = base.ripe;
-}
+const base = require("./base");
+require("./ripe");
+const ripe = base.ripe;
 
 /**
  * Verifies if the provided tag exists in the sequence of