+++ src/js/base/logic.js
@@ -1,20 +1,8 @@
-if (
-    typeof require !== "undefined" &&
-    (typeof window === "undefined" ||
-        // eslint-disable-next-line camelcase
-        typeof __webpack_require__ !== "undefined" ||
-        (typeof navigator !== "undefined" && navigator.product === "ReactNative"))
-) {
-    // eslint-disable-next-line no-redeclare,no-var
-    var base = require("./base");
-    // eslint-disable-next-line no-var
-    var locales = require("../locales");
-    require("./ripe");
-    // eslint-disable-next-line no-redeclare,no-var
-    var ripe = base.ripe;
-    // eslint-disable-next-line no-var
-    var LOCALES_BASE = locales.LOCALES_BASE;
-}
+const base = require("./base");
+const locales = require("../locales");
+require("./ripe");
+const ripe = base.ripe;
+const LOCALES_BASE = locales.LOCALES_BASE;
 
 /**
  * Adds a new bundle object to the currently set of registered